// backend/src/common/interfaces/google-profile.interface.ts
export interface GoogleProfile {
  id: string;
  displayName: string;
  name?: {
    familyName?: string;
    givenName?: string;
  };
  emails?: {
    value: string;
    verified?: boolean;
  }[];
  photos?: {
    value: string;
  }[];
  provider: "google";
  _raw?: string;
  _json: {
    sub: string;
    name: string;
    given_name?: string;
    picture?: string;
    email?: string;
    email_verified?: boolean;
  };
}

export interface UserFromGoogle {
  provider: "GOOGLE";
  providerId: string;
  name: string;
  email: string;
}